app/pyproject.toml
[tool.poetry]
name = "mannequins-api"
version = "0.1.0"
description = ""
authors = ["Ato Toffah <atotoffah@gmail.com>"]
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.1"
uvicorn = "^0.27.0"
pydantic-settings = "^2.4.0"
python-multipart = "^0.0.6"
pymongo = "^4.3.3"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "4.0.1"
fastapi-pagination = "^0.12.26"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

